requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic>=2.11,<3.0
numpy>=2.3,<3.0
pandas>=2.3,<3.0
pyarrow>=21.0,<22.0
scikit-learn>=1.7,<2.0
joblib>=1.5,<2.0
faiss-cpu>=1.12,<2.0
torch>=2.10,<3.0
transformers>=4.57,<5.0
httpx>=0.28,<1.0
